Yeni klasör/requirements.txt
PyQt6
pyqtgraph
pyserial

